tests/requirements.txt
django-subcommand
